feat: 完成排序功能

+++ src/StrongBuy.Blazor/Models/Product.cs
@@ -4,8 +4,11 @@ public class Product
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
+
     public string Description { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+
+    // public decimal Price { get; set; }
+    public double Price { get; set; }
     public string Category { get; set; } = string.Empty;
     public List<string> Subcategories { get; set; } = new();
     public string Brand { get; set; } = string.Empty;
